==> Spotify Hit Prediction/1_Preprocessing/Data Wrangling/2_Spotipy - Add Song URIs.ipynb <==
In this Notebook I will add the Spotify track URIs to the billboards data set.

I will use the spotipy package because the querys are less sensitive and easier to create than connecting directly to the Spotify API.

Methodology:

Import packages & .env to store API keys Easily done thanks to the lessons learnt from the API mini project. DONE
Authenticate myself via the client credentials through the spotipy package After reading through the documentation this methd of connecting to the API seems easier and fewere errors with the query portion of the GET request.
Create a funtion to get URIs for songs from the JSON output of the API This is proving to be very dificult. There are many songs that are not
Apply the funtion to the Billboards dataset
Check for missing values, duplicates, etc. and make changes to the function as required

# get api key from your .env file

import os

from dotenv import load_dotenv

from pathlib import Path

 

dotenv_path = Path('./SpotifyAPIKey.env')

load_dotenv(dotenv_path=dotenv_path)

 

#load_dotenv()

API_CLIENT_ID = os.getenv('SPOTIFY_CLIENT_ID')
API_CLIENT_SECRET = os.getenv('SPOTIFY_CLIENT_SECRET')
 

    

import pandas as pd
import spotipy
from spotipy.oauth2 import SpotifyClientCredentials

#https://spotipy.readthedocs.io/en/2.19.0/

#Create a function to get the URI for a Title and Artist

client_credentials_manager = SpotifyClientCredentials(client_id=API_CLIENT_ID,client_secret=API_CLIENT_SECRET)
sp = spotipy.Spotify(client_credentials_manager=client_credentials_manager)

def URI(titleName, artistName):
    track = sp.search(q="artist:" + f"{artistName}"+" track:"+f"{titleName}",type='track',limit=50)
    result = track['tracks']['items']
    result = sorted(result, key = lambda i: i['popularity'], reverse = True)
    
    for items in result:
        for artist in items['artists']:
            if artist['name'].lower() == artistName.lower():
                uri = items['id']
                return(uri)
            elif artist['name'].lower() in artistName.lower():
                uri = items['id']
                return uri
            
    track = sp.search(titleName,type='track',limit=50)
    result = track['tracks']['items']
    result = sorted(result, key = lambda i: i['popularity'], reverse = True)
    
    for items in result:
        for artist in items['artists']:
            if artist['name'].lower() == artistName.lower():
                uri = items['id']
                return uri

    for items in result:
        for artist in items['artists']:
            if artist['name'].lower() in artistName.lower():
                uri = items['id']
                return uri
            
    track = sp.search(titleName,type='track',limit=50)
    result = track['tracks']['items']
    result = sorted(result, key = lambda i: i['popularity'], reverse = True)
    
    for items in result:
        if items['name'].lower() == titleName.lower():
            uri = items['id']
            return uri
        elif items['name'].lower() in titleName.lower():
            uri = items['id']
            return uri
    

print(URI('caroline', "amine"))

print(URI("the way you love me", "faith hill"))

#Import the Billboard Data
df=pd.read_csv('BillboardFullDataset.csv')

#As discovered there are some issues with some of the data. They contain \xa0 in place of spaces. I change those here. 

df['Title'] = df['Title'].apply(lambda x: str(x).replace(u'\xa0', u' '))
df['Artist'] = df['Artist'].apply(lambda x: str(x).replace(u'\xa0', u' '))

df.head(10)

df = df[['Title','Artist']]

df['Title'] = df['Title'].str.lower()
df['Artist'] = df['Artist'].str.lower()

df.head()

def append_URIs(dataframe):
    for lab, row in dataframe.iterrows():
        dataframe.loc[lab, 'URI'] = URI(row['Title'],row['Artist'])

append_URIs(df)

df.describe()

There are issues here. There are missing URIs and duplicate URIs.

I'll first attack the missing URIs - I think these can be solved by changing the functions slightly to make sure there are no name issues e.g. sounfs featuring multiple artist on billboard but only showing the main artist on spotify

Round 1: Stared with 151 missing URIs.

Round 2: After fixing 1 Artist name I am down to 150 missing URIs.

Round 3: I changed my append_URIs function to take in all lowercase values for artist. I am now down to missing 138 URIs

Round 4: I added some conditions to my function to also look through track names and match the artist and from there. I also added a condition to look at lowercase songs. I am down to 14 missing URIs.

Note: this is weird.. function to get the uri works on some of these missing values so there must be an issue with the append uri's function.

missing = df.loc[df['URI'].isna()]

missing

#Dropping the last two NaN values.

df = df.dropna()

df_dups = df[df.duplicated('URI', keep=False)]

df_dups.sort_values(by='URI', ascending=True)

df.drop_duplicates(subset = 'URI', inplace=True)

df.info()

df.to_csv('BillboardWithURI.csv')
